aws: Search all cluster tags, even if legacy tag doesn't match
hasClusterTag incorrectly returns early when it finds the ClusterTag on an AWS
resource, even when the ClusterID does not match the configured
ClusterID. This results in the method not searching for the non-legacy
tag, if the legacy tag comes first in the list of tags.

This patch fixes the logic to only return when the tag value matches the
configured clusterID.

This may fix other issues, but currently this fixes the case where a
subnet is tagged with both legacy `ClusterName` and the new
`kubernetes.io/cluster/$CLUSTER_NAME` tags. If the legacy tag comes
before the new tag it would return before even looking as the new tags,
which prevents the correct subnets from being attached to the ELB.